+++ my-react-app/src/App.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import './style.css';
+import { BrowserRouter } from 'react-router-dom';
 import Routing from './components/Routing';
+import Header from './components/Header';
 
 function App() {
   return (
-    <div className="App">
+    <BrowserRouter>
+      <Header />
       <Routing />
-    </div>
+    </BrowserRouter>
   );
 }
