+++ my-react-app/src/components/Routing.tsx
@@ -1,21 +1,16 @@
+import AboutPage from '../pages/AboutPage';
+import MainPage from '../pages/MainPage';
+import NotFoundPage from '../pages/NotFoundPage';
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import AboutPage from './AboutPage';
-import Header from './Header';
-import MainPage from './MainPage';
-import NotFoundPage from './NotFoundPage';
 
 function Routing() {
   return (
-    <Router>
-      <Header />
-
       <Routes>
         <Route path="/main" element={<MainPage />} />
         <Route path="/about" element={<AboutPage />} />
         <Route path="*" element={<NotFoundPage />} />
       </Routes>
-    </Router>
   );
 }
