+++ my-react-app/src/style.css
@@ -3,7 +3,6 @@
 body {
   background: #F5F5F5;
   font-family: 'Croissant One', cursive;
-  transition: all 0.05s ease-in-out;
   /* font-family: 'Lato', sans-serif; */
 }
